pyproject.toml
[project]
name = "render-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.9",
    "jinja2>=3.1.6",
    "libsql-experimental>=0.0.40",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.38.0",
]
